requirements.txt
Pillow>=10.0
torch>=2.0
openai>=1.0
python-dotenv>=1.0
numpy>=1.24
